+++ app/platform.js
@@ -1,5 +1,5 @@
-document.body.classList.add(window.process.platform())
-if (window.process.windowsStore()) {
+document.body.classList.add(window.runtime.platform())
+if (window.runtime.windowsStore()) {
   document.body.classList.add('store')
 }
 export default {}