+++ app/process-renderer.js
@@ -11,8 +11,8 @@ window.onload = async (e) => {
   window.stretchly.onShowNotification(async (text, silent) => {
     __electronLog.info(`Stretchly: showing notification "${text}" (silent: ${silent})`)
     const title = await window.utils.shouldShowNotificationTitle(
-      await window.process.platform(),
-      await window.process.getSystemVersion()
+      await window.runtime.platform(),
+      await window.runtime.getSystemVersion()
     )
       ? 'Stretchly'
       : ''
@@ -51,7 +51,7 @@ window.onload = async (e) => {
   })
 
   async function notifyNewVersion (silent) {
-    const title = await window.utils.shouldShowNotificationTitle(await window.process.platform(), await window.process.getSystemVersion()) ? 'Stretchly' : ''
+    const title = await window.utils.shouldShowNotificationTitle(await window.runtime.platform(), await window.runtime.getSystemVersion()) ? 'Stretchly' : ''
     const notification = new Notification(title, {
       body: await window.i18next.t('process.newVersionAvailable'),
       silent,