+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- hide autostart option for Windows store
+
+### Changed
+- hide update features in Windows Store, Snap, Flatpak versions by default
+
 ## [1.18.1] - 2025-9-27
 ### Fixed
 - fix Linux crashes on app start